Allow fields with more than 255 chars
The max_value on each form field type is used for max length on all text
types and for max amount of selections on multi choice fields. It was 
limited to 255 due to the previous way how submitted form data was
saved. But that limit does not make sense for a TextArea, which should
not have a limit.